v1.31
RELEASE NOTES
=============

v1.31 - 6 Sept 29 2014

BUG: Poll all friends was only polling 1 friend.
FIX: Removed a debug statement that was causing the problem.

BUG: Center window sometimes comes up grey.
CAUSE: Seems to coincide at least in some cases with VideoRecorder
being asked for while in background. I got an alert with that message
once coincident with the center screen being grey.
SOLUTION: Re-architect the videoRecorder so that it can be initialized
and disposed in the appropriate points in the app lifecycle. See
ReleaseNotesDetails for this release.

BUG: When app has finished background time and you get a notification.
It says downloading but hangs there. This bug was fixed by requesting
background time in didReceiveNotification. But it was reintroduced with
due to another cause by the instrumentation I added to elucidate the
grey center window bug. See how this happened in ReleaseNotesDetails.
SOLUTION: Fix for grey center window should fix this as well.