v1.34
=============
v1.34 - 8 - 9/10/2014

BUG:  Downloads get stuck when in Cellular.
CAUSE: It turns out that ios reserves the option not to initiate your
downloads in cellular if they were initiated in the background as by
remote notification. The problem is that it also doesn't initiate them
for you when the app moves to the foreground.
FIX: Added methods to get transferInfo (from NSTask) as well as obTask
info from FileTransferManger. In AppSync handleStuckDownloads handles
all anomalies for videos expected to be downloading. Including
  - Stalled download due to initiation in background while on cellular
network.
  - Loss of NSTask under ObTask due to termination while download was
in progress.
  - Loss of OBTask which should never happen but which I believe I have
seen once or twice.
See release details for more info.

BUG: Grey center box where camera preview should be. Happens
occasionally.
CAUSE: Multifactorial. See release details.
FIX:
 - Re architected the VideoRecorder and adjusted the states where we
instantiate and destroy. See release details.
 - Added an indicator for our debugging purposes to show when we are
waiting for the system to give us the camera. If we get a
AVCaptureSessionRuntimeException as we sometimes do and the system asks
us to try again later I indicate the number of attempts in this debug
indicator.
POSSIBLE KNOWN BUG: This fixe seems to have fixed almost all instances
of the problem. I believe I have seen a grey screen though with no
debugging indicator showing. But I cant readily reproduce.

BUG: After downloading or uploading a video the app occasionally hangs
on startup.
CAUSE: With the move to the new FileTransferManager were violating the
concurrency contract we have with NSManagedObjects. You may not delete,
fetch, or pass managed objects from one thread to another. They must
always be handled on the thread that the NSManagedObjectContext that
was used to fetch them was created.
FIX:
  - Switch to the new Apple NSManagedObject architecture where we
specify a concurrency type (NSMainQueueConcurrencyType)
  - Ensure that calls in specifically events from FTM that are not
running on the main thread are dispatched to the main thread.

ARCHITECTURE:
  - Boot should be a category on appDelegate not on homeViewController.
Refactored.
  - Not calling many of our initialization methods on the optimal
lifecycle events. See release details for changes.