v1.40
RELEASE NOTES
=============
AdHoc Distribution Release:

v1.40 - 10 - 9/12/2014

FEATURE: Integrate with Testflight so that we can distribute and update
the ios version of the app to alpha users over the air.

BUG: When I load the ipa from TestFlight and run on the I4 I get a
crash on startup after shutting the app down in the task manager.
CAUSE: In VideoRecorder init -  I was passing a pointer to an error
object as a parameter to a method. The method was filling it in several
cases. In one case it was filling on an async thread that was
dispatched by the method. In the case the async thread completed before
the method did everything was fine which expalains the race and why it
worked on one phone and not the other. Pretty obvious mistake in
hindsight.
SOLUTION: Create a local block variable in the block to handle.

BUG: Push_token not being sent on boot when we go through a new
registration.
FIX: Refactor the boot code.

BUG: Phone is not receiving notification when built as a release target
and loaded over the air from TestFlight.
CAUSE: Apple requires a production APNS certificate as well as hitting
the production gateway to send notifications to phones with the app
distributed as an ad_hoc release.
SOLUTION: Refactor the APNS gem so that we can hit either production or
dev gateway at apple with production or dev pem on a pre request basis.
Update the server to accommodate the new architecture. Update the app
to send a device_build = prod, dev depending on whether it is an xcode
dev build or a production release distributed by TestFlight.