v1.41
RELEASE NOTES
=============

v1.41 - 11 - 9/16/2014

BUG: Request background when we handle a notification via
didFinishLaunching with options as well as
DidReceiveRemoteNotification. This is probably not necessary but it
should not hurt.

BUG: HomeViewController is trying to initialize the camera when app is
launched in the background due to notifications.
FIX: Change logic for instantiation of VideoRecorder in
HomeViewController. See Release Details.

BUG: When a video fails permanently we keep reinstating it. We should
treat it as viewed and let it be killed in the next download cycle.
FIX: Change friend.deleteAllViewed to deleteAllFailedOrViewed

BUG: When we force requeue of a video we should not create a new video
object but just requeue the existing one.
FIX: Change logic in appSync.queueForDownload